--- tests/unit/mod.rs
@@ -1,2 +1,0 @@
-mod config;
-mod ui;
